tech: moved rest of async methods
